apply bable stream changes for alpaka2
- performance is good

Dot kernel consits a workaround because float accumulation in a row
fails after 2k elements. This is an issue of the verification itself. By
using more threads or on CPU use all frames and the frame extent this
issue is solved because each starting point is doing less sequential
accumulations.